ENH: Update the internal index of TemplateFlow
This PR introduces `templateflow.conf.update()` to allow for the update
the TemplateFlow structure pointed by TEMPLATEFLOW_HOME.

It has a "local" operation mode (which uses the skeleton distributed
within the installed package) to ensure the continuity of previous work
on this line. This means that, when re-installing templateflow, the
TEMPLATEFLOW_HOME will still be updated the same way it used to be.

This PR is limited, as it does not allow to point templateflow to a
particular release of the archive. The main reason for that is that we
would need to have a version controlled repository of skeletons (which
does not exist yet).

Resolves: #37