fix: new approach to multiversioned docs
This PR relies on a little javascript to create the dropdown menu to
select the version.
The sphinxcontrib-versioning package is not anymore used, and the
different versions of documentation are just accumulated on the
`gh-pages` branch of the repo.
Therefore, all versions are not built anymore with every build.
The new approach completely decouples the generation of versions
of documentation from the chooser, allowing us flexibility to change
themes or completely change the structure, so long the new theme or
documentation scheme is capable of interpreting the `versions.json` file
at the root of `gh-pages`.